Extracted test class to work with gz
